pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "py-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
db = ["psycopg2-binary>=2.9"]
excel = ["pandas>=2.3", "openpyxl>=3.1"]
cipher = ["ff3>=1.0"]
all = ["py-template[db,excel,cipher]"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-html>=4.0",
    "pytest-cov>=6.0",
    "allure-pytest>=2.13",
    "ruff>=0.11",
    "mypy>=1.10",
    "types-pyyaml>=6.0",
    "pre-commit>=4.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["common", "adapters"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "gs_req(req_id): GS 체크리스트 항목 ID (예: GS-07, GS-31). 구분은 항목으로 조회.",
    "isms(id): ISMS-P 체크리스트 항목 ID",
    "integration: 통합 테스트 (DB, 파일 등 외부 자원 필요)",
    "e2e: 종단 간 테스트 (실제 사용자 시나리오)",
    "local_only: 로컬에서만 실행 (CI에서는 제외, pytest -m \"not local_only\")",
]

[tool.coverage.run]
source = ["src"]
omit = ["src/*/__init__.py"]

[tool.coverage.report]
show_missing = true
fail_under = 0

[tool.coverage.html]
directory = "reports/coverage"
